create new codeEditor
